src/robot/requirements.txt
argparse==1.2.1
distribute==0.6.24
jsonrpclib==0.1.6
psutil==3.2.2
pyOSC==0.3.5b-5294
pyserial==2.7
wsgiref==0.1.2
